+++ app/src/main/java/com/example/jianming/beans/DLFilePathBean.kt
@@ -2,5 +2,5 @@ package com.example.jianming.beans
 
 import java.io.File
 
-class DLFilePathBean(var index: Int, var src: String, var dest: File, var picIndex: Int, var position: Int, var encrypted: Boolean) {
+class DLFilePathBean(var index: Long, var src: String, var dest: File, var picIndex: Int, var position: Int, var encrypted: Boolean) {
 }
